switch to uint64_t everywhere (also fixes mac os compilation)

+++ dorado/read_pipeline/DuplexSplitNode.h
@@ -15,17 +15,17 @@ struct DuplexSplitSettings {
     bool enabled = true;
     bool simplex_mode = false;
     float pore_thr = 2.2;
-    size_t pore_cl_dist = 4000;  // TODO maybe use frequency * 1sec here?
+    uint64_t pore_cl_dist = 4000;  // TODO maybe use frequency * 1sec here?
     //maximal 'open pore' region to consider (bp)
-    size_t max_pore_region = 500;
+    uint64_t max_pore_region = 500;
     //usually template read region to the left of potential spacer region
-    size_t strand_end_flank = 1200;
+    uint64_t strand_end_flank = 1200;
     //trim potentially erroneous (and/or PCR adapter) bases at end of query
-    size_t strand_end_trim = 200;
+    uint64_t strand_end_trim = 200;
     //adjusted to adapter presense and potential loss of bases on query, leading to 'shift'
-    size_t strand_start_flank = 1700;
+    uint64_t strand_start_flank = 1700;
     //minimal query size to consider in "short read" case
-    size_t min_flank = 300;
+    uint64_t min_flank = 300;
     float flank_err = 0.22;
     float relaxed_flank_err = 0.275;
     int adapter_edist = 4;
@@ -35,7 +35,7 @@ struct DuplexSplitSettings {
     uint64_t expect_adapter_prefix = 200;
     //in samples
     uint64_t expect_pore_prefix = 5000;
-    int middle_adapter_search_span = 1000;
+    uint64_t middle_adapter_search_span = 1000;
     float middle_adapter_search_frac = 0.2;
 
     //TODO put in config
